schedule tasks
WPKG Service start type is set to Auto which doesn't actually ever run on Win 10, change it to Auto (Delayed)
Schedule the Service to run every day to cope with users who leave their computer on.
Schedule the WPKGsync task to run every hour so we can have a chance to sync the data before the next run.